search/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
pydantic-settings
rapidfuzz
python-Levenshtein